requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
progress==1.6
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
tqdm==4.64.0
urllib3==1.26.9
yadisk==1.2.15
